+++ docs/history.md
@@ -3,8 +3,10 @@
 <!-- cspell:ignore Adrayon,cantgirl,elgirs,Fallenfor,Finalkids,Hispa,Kaethela,lenamek,mgln,Moucheron,Quipet,Sanio,sisuka,SuperpaperChildy,Toooschi -->
 <!-- markdownlint-disable MD001 -->
 
-### _v1.25.14_ - October 3rd, 2023
+### _v1.26.0_ - October 3rd, 2023
 
+- The following babies have new effects:
+  - Worry Baby (#167) - "Touching items/pickups causes teleportation" --> "Touching items/pickups spawns Mega Troll Bombs"
 - Fixed bugs with the following babies:
   - Love Eye Baby (#249) - No longer replaces Swingers. (Thanks Gamonymous)